+++ lib/crass/parser.rb
@@ -79,6 +79,17 @@ module Crass
         next if node.nil?
 
         case node[:node]
+        when :at_rule
+          string << '@'
+          string << node[:name]
+          string << self.stringify(node[:prelude], options)
+
+          if node[:block]
+            string << '{' << self.stringify(node[:block], options) << '}'
+          else
+            string << ';'
+          end
+
         when :comment
           string << node[:raw] unless options[:exclude_comments]